Use HeightmapColorLookupTable as stub for generic table
Need to support both int-Color and int-float mappings. Create a generic class so that it is easier to implement SpeedLookupTable soon
