+++ Scripts/HeightmapLookupTable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12305298a2c974e4fb5ca6d4e6d29030
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
